+++ HISTORY.rst
@@ -5,10 +5,17 @@ History
 1.0.2 (unreleased)
 ------------------
 
+- Rename ``usbid.device.devicelist`` to ``usbid.device.device_list``. B/C
+  import available.
+  [rnix, 2014-02-26]
+
+- Introduce ``usbid.device.device_by_path`` utility.
+  [rnix, 2014-02-26]
+
 - use latest ``usb.id`` file version 2014.02.03 from 
   http://www.linux-usb.org/usb.ids
   [jensens, 2014-02-20]
-i
+
 
 1.0.1
 -----